requirements.txt
eel
faster-whisper
numpy
pyyaml
transformers
sacremoses
sentencepiece
#sounddevice
#soundcard
soundfile
torch
torchvision
torchaudio
